# [SYSTEMDS-3444] Spark Write Compressed