--- a/src/main/java/org/apache/sysds/runtime/compress/colgroup/ColGroupSDCSingle.java
+++ b/src/main/java/org/apache/sysds/runtime/compress/colgroup/ColGroupSDCSingle.java
@@ -582,6 +582,11 @@ public class ColGroupSDCSingle extends ASDC {
 	}
 
 	@Override
+	public AColGroup appendNInternal(AColGroup[] g) {
+		return null;
+	}
+
+	@Override
 	public String toString() {
 		StringBuilder sb = new StringBuilder();
 		sb.append(super.toString());