--- a/src/main/java/org/apache/sysds/runtime/compress/colgroup/ColGroupSDCSingleZeros.java
+++ b/src/main/java/org/apache/sysds/runtime/compress/colgroup/ColGroupSDCSingleZeros.java
@@ -815,6 +815,11 @@ public class ColGroupSDCSingleZeros extends ASDCZero {
 	}
 
 	@Override
+	public AColGroup appendNInternal(AColGroup[] g) {
+		return null;
+	}
+
+	@Override
 	public String toString() {
 		StringBuilder sb = new StringBuilder();
 		sb.append(super.toString());