--- a/src/main/java/org/apache/sysds/runtime/compress/colgroup/ColGroupSDCZeros.java
+++ b/src/main/java/org/apache/sysds/runtime/compress/colgroup/ColGroupSDCZeros.java
@@ -717,7 +717,7 @@ public class ColGroupSDCZeros extends ASDCZero {
 
 	@Override
 	protected AColGroup copyAndSet(int[] colIndexes, ADictionary newDictionary) {
-		return create(colIndexes, _numRows, newDictionary, _indexes, _data, getCounts());
+		return create(colIndexes, _numRows, newDictionary, _indexes, _data, getCachedCounts());
 	}
 
 	@Override