--- a/src/main/java/org/apache/sysds/runtime/compress/colgroup/ColGroupUncompressed.java
+++ b/src/main/java/org/apache/sysds/runtime/compress/colgroup/ColGroupUncompressed.java
@@ -776,6 +776,11 @@ public class ColGroupUncompressed extends AColGroup {
 	}
 
 	@Override
+	public AColGroup appendNInternal(AColGroup[] g) {
+		return null;
+	}
+
+	@Override
 	public String toString() {
 		StringBuilder sb = new StringBuilder();
 		sb.append(super.toString());