… src/main/java/org/apache/sysds/runtime/compress/colgroup/mapping/MapToCharPByte.java
… src/main/java/org/apache/sysds/runtime/compress/colgroup/mapping/MapToInt.java
… src/main/java/org/apache/sysds/runtime/compress/colgroup/mapping/MapToZero.java
… src/main/java/org/apache/sysds/runtime/compress/colgroup/offset/AOffset.java
… src/main/java/org/apache/sysds/runtime/compress/io/CompressWrap.java
… src/main/java/org/apache/sysds/runtime/compress/io/CompressedWriteBlock.java
… src/main/java/org/apache/sysds/runtime/compress/io/ReaderCompressed.java
… src/main/java/org/apache/sysds/runtime/compress/io/WriterCompressed.java
… src/main/java/org/apache/sysds/runtime/compress/lib/CLALibCombine.java
… src/main/java/org/apache/sysds/runtime/compress/lib/CLALibDecompress.java
… src/main/java/org/apache/sysds/runtime/compress/lib/CLALibSlice.java
… src/main/java/org/apache/sysds/runtime/compress/utils/DoubleCountHashMap.java
… src/main/java/org/apache/sysds/runtime/instructions/cp/DataGenCPInstruction.java
… src/main/java/org/apache/sysds/runtime/instructions/spark/WriteSPInstruction.java
… src/main/java/org/apache/sysds/runtime/instructions/spark/functions/ExtractBlockForBinaryReblock.java
… src/main/java/org/apache/sysds/runtime/instructions/spark/utils/RDDAggregateUtils.java
… src/main/java/org/apache/sysds/runtime/matrix/data/MatrixBlock.java
… src/test/java/org/apache/sysds/test/component/compress/colgroup/ColGroupNegativeTests.java
… src/test/java/org/apache/sysds/test/component/compress/combine/CombineTest.java
… src/test/java/org/apache/sysds/test/component/compress/io/IOCompressionTestUtils.java
… src/test/java/org/apache/sysds/test/component/compress/io/IOEmpty.java
… src/test/java/org/apache/sysds/test/component/compress/io/IOSpark.java
… src/test/java/org/apache/sysds/test/component/compress/io/IOTest.java
… src/test/java/org/apache/sysds/test/component/compress/mapping/MappingTests.java
… src/test/java/org/apache/sysds/test/functions/io/compressed/WriteCompressedTest.java
… src/test/scripts/functions/io/compressed/WriteCompressedTest.dml