Recent commit did not address the spark writing completely therefore this PR follows up and adds the writing of a RDD to disk.