@Baunsgaard 2022-10-17
Preview Compressed format for binary matrices, comparing text binary and compressed.:
![image](https://user-images.githubusercontent.com/9947148/196168405-f059552f-d959-4c12-9779-7d51b7ac891f.png)