@Baunsgaard 2022-10-17
Reading -> plus -> sum of ... empty ... 256k x 2k matrix.

Binary:
```
1.536E9
SystemDS Statistics:
Total elapsed time:		5.827 sec.
Total compilation time:		0.383 sec.
Total execution time:		5.444 sec.
Cache hits (Mem/Li/WB/FS/HDFS):	1/0/0/0/1.
Cache writes (Li/WB/FS/HDFS):	0/0/1/0.
Cache times (ACQr/m, RLS, EXP):	0.076/0.000/3.880/0.000 sec.
HOP DAGs recompiled (PRED, SB):	0/0.
HOP DAGs recompile time:	0.000 sec.
Total JIT compile time:		0.253 sec.
Total JVM GC count:		0.
Total JVM GC time:		0.0 sec.
Heavy hitter instructions:
 #  Instruction  Time(s)  Count
 1  +              5.190      1
 2  uak+           0.246      1
 3  createvar      0.006      2
 4  print          0.001      1
 5  rmvar          0.001      2
```

Compressed:

```
1.536E9
SystemDS Statistics:
Total elapsed time:		0.437 sec.
Total compilation time:		0.376 sec.
Total execution time:		0.060 sec.
Cache hits (Mem/Li/WB/FS/HDFS):	1/0/0/0/1.
Cache writes (Li/WB/FS/HDFS):	0/0/0/0.
Cache times (ACQr/m, RLS, EXP):	0.045/0.000/0.000/0.000 sec.
HOP DAGs recompiled (PRED, SB):	0/0.
HOP DAGs recompile time:	0.000 sec.
Total JIT compile time:		0.132 sec.
Total JVM GC count:		0.
Total JVM GC time:		0.0 sec.
Heavy hitter instructions:
 #  Instruction  Time(s)  Count
 1  +              0.050      1
 2  createvar      0.006      2
 3  uak+           0.003      1
 4  print          0.001      1
 5  rmvar          0.000      2
```