@Baunsgaard 2022-10-18
Initial Results:
Writing a Matrix Block of 64k x 2k.

![Screenshot from 2022-10-18 18-27-38](https://user-images.githubusercontent.com/9947148/196488768-a4a62634-0124-4aaa-bbc3-65f9a119dccc.png)
![Screenshot from 2022-10-18 18-27-58](https://user-images.githubusercontent.com/9947148/196488781-64e703fd-b21b-429d-82f9-184430810482.png)