@Baunsgaard 2022-10-19
Optimization 1. Do not use Hadoop compression but only internal serialization:
Improved performance
Read: 1.6 -> 0.939 sec 1k blocks and 0.9 to 0.6 with 128k blocks.
Write: 2 -> 1.4 sec

![image](https://user-images.githubusercontent.com/9947148/196721194-fb21a77b-00d1-4f42-84db-dd1f2c489a8a.png)